Rename backtrace framework
Missed one more conflicting framework
