+++ src/mca/prtebacktrace/base/Makefile.am
@@ -9,7 +9,7 @@
 #                         University of Stuttgart.  All rights reserved.
 # Copyright (c) 2004-2005 The Regents of the University of California.
 #                         All rights reserved.
-# Copyright (c) 2019      Intel, Inc.  All rights reserved.
+# Copyright (c) 2019-2020 Intel, Inc.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -17,8 +17,8 @@
 # $HEADER$
 #
 
-noinst_LTLIBRARIES = libmca_backtrace_execinfo.la
+headers += \
+        base/base.h
 
-libmca_backtrace_execinfo_la_SOURCES = \
-    backtrace_execinfo.c \
-    backtrace_execinfo_component.c
+libmca_prtebacktrace_la_SOURCES += \
+	base/backtrace_component.c
