+++ src/mca/prtebacktrace/execinfo/Makefile.am
@@ -9,8 +9,7 @@
 #                         University of Stuttgart.  All rights reserved.
 # Copyright (c) 2004-2005 The Regents of the University of California.
 #                         All rights reserved.
-# Copyright (c) 2010 Cisco Systems, Inc.  All rights reserved.
-# Copyright (c) 2019      Intel, Inc.  All rights reserved.
+# Copyright (c) 2019-2020 Intel, Inc.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -18,21 +17,8 @@
 # $HEADER$
 #
 
-# main library setup
-noinst_LTLIBRARIES = libmca_backtrace.la
-libmca_backtrace_la_SOURCES =
+noinst_LTLIBRARIES = libmca_prtebacktrace_execinfo.la
 
-# local files
-headers = backtrace.h
-libmca_backtrace_la_SOURCES += $(headers)
-
-# Conditionally install the header files
-if WANT_INSTALL_HEADERS
-prrtedir = $(prrteincludedir)/$(subdir)
-nobase_prrte_HEADERS = $(headers)
-endif
-
-include base/Makefile.am
-
-distclean-local:
-	rm -f base/static-components.h
+libmca_prtebacktrace_execinfo_la_SOURCES = \
+    backtrace_execinfo.c \
+    backtrace_execinfo_component.c
