+++ src/mca/prtebacktrace/none/backtrace_none_component.c
@@ -12,7 +12,7 @@
  *                         All rights reserved.
  * Copyright (c) 2015      Los Alamos National Security, LLC. All rights
  *                         reserved.
- * Copyright (c) 2019      Intel, Inc.  All rights reserved.
+ * Copyright (c) 2019-2020 Intel, Inc.  All rights reserved.
  * Copyright (c) 2019      Research Organization for Information Science
  *                         and Technology (RIST).  All rights reserved.
  * $COPYRIGHT$
@@ -24,13 +24,13 @@
 
 #include "prrte_config.h"
 
-#include "src/mca/backtrace/backtrace.h"
+#include "src/mca/prtebacktrace/prtebacktrace.h"
 
 BEGIN_C_DECLS
     PRRTE_EXPORT extern const prrte_backtrace_base_component_2_0_0_t prrte_backtrace_none_component;
 END_C_DECLS
 
-const prrte_backtrace_base_component_2_0_0_t prrte_backtrace_none_component = {
+const prrte_backtrace_base_component_2_0_0_t prrte_prtebacktrace_none_component = {
     /* First, the mca_component_t struct containing meta information
        about the component itself */
     .backtracec_version = {
