+++ src/mca/prtebacktrace/none/configure.m4
@@ -1,3 +1,4 @@
+# -*- shell-script -*-
 #
 # Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
 #                         University Research and Technology
@@ -9,16 +10,26 @@
 #                         University of Stuttgart.  All rights reserved.
 # Copyright (c) 2004-2005 The Regents of the University of California.
 #                         All rights reserved.
-# Copyright (c) 2019      Intel, Inc.  All rights reserved.
+# Copyright (c) 2010      Cisco Systems, Inc.  All rights reserved.
+# Copyright (c) 2019-2020 Intel, Inc.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
 #
 # $HEADER$
 #
+AC_DEFUN([MCA_prrte_prtebacktrace_none_PRIORITY], [0])
 
-headers += \
-        base/base.h
+AC_DEFUN([MCA_prrte_prtebacktrace_none_COMPILE_MODE], [
+    AC_MSG_CHECKING([for MCA component $1:$2 compile mode])
+    $3="static"
+    AC_MSG_RESULT([$$3])
+])
 
-libmca_backtrace_la_SOURCES += \
-	base/backtrace_component.c
+
+# MCA_prtebacktrace_none_CONFIG(action-if-can-compile,
+#                        [action-if-cant-compile])
+# ------------------------------------------------
+AC_DEFUN([MCA_prrte_prtebacktrace_none_CONFIG],[
+    AC_CONFIG_FILES([src/mca/prtebacktrace/none/Makefile])
+])
