+++ src/mca/prtebacktrace/printstack/backtrace_printstack_component.c
@@ -12,7 +12,7 @@
  *                         All rights reserved.
  * Copyright (c) 2015      Los Alamos National Security, LLC. All rights
  *                         reserved.
- * Copyright (c) 2019      Intel, Inc.  All rights reserved.
+ * Copyright (c) 2019-2020 Intel, Inc.  All rights reserved.
  * Copyright (c) 2019      Research Organization for Information Science
  *                         and Technology (RIST).  All rights reserved.
  * $COPYRIGHT$
@@ -24,16 +24,16 @@
 
 #include "prrte_config.h"
 
-#include "src/mca/backtrace/backtrace.h"
+#include "src/mca/prtebacktrace/prtebacktrace.h"
 
-const prrte_backtrace_base_component_2_0_0_t prrte_backtrace_execinfo_component = {
+const prrte_backtrace_base_component_2_0_0_t prrte_prtebacktrace_printstack_component = {
     /* First, the mca_component_t struct containing meta information
        about the component itself */
     .backtracec_version = {
         PRRTE_BACKTRACE_BASE_VERSION_2_0_0,
 
         /* Component name and version */
-        .mca_component_name = "execinfo",
+        .mca_component_name = "printstack",
         PRRTE_MCA_BASE_MAKE_VERSION(component, PRRTE_MAJOR_VERSION, PRRTE_MINOR_VERSION,
                                     PRRTE_RELEASE_VERSION),
     },
